move get_related_cases to bihar custom app
